multistack/requirements.txt
aws-cdk.core
aws_cdk.aws_ec2
aws_cdk.aws_iam
aws_cdk.aws_logs
aws_cdk.aws_autoscaling
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_elasticloadbalancingv2_targets
aws_cdk.aws_cloudwatch
aws-cdk.aws-rds
aws_cdk.aws_ssm
-e .
